requirements.txt
asgiref==3.8.1
Django==5.2.1
django-admin-interface==0.30.0
django-colorfield==0.14.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-slugify==8.0.4
qrcode==8.2
sqlparse==0.5.3
text-unidecode==1.3
